1.6.0: support for actions artifact-v4, Nix docker builds (#65)
This fixes a breaking change in GHA's upload-artifact v4, which CI
relied upon in order to merge Docker manifests.

To make the CI easier for me to maintain, the Dockerfile-based builds
are gone in favour of a single Nix-based container build. This has the
breaking change of dropping support for 32-bit ARM, and the benefit of
bringing reproducible builds to container images. Given the armv6 images
had practically no downloads, few if any existing users should be
accepted (do please open an issue if this is you!).

This in turn makes leng practically a `FROM SCRACH` image, which reduces
the image attack surface. Image size remains about the same (pretty
small at 7MB).

Because of the breaking changes above, this makes release 1.6